kserve/vision/requirements.txt
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudevents==1.9.0
cmake==3.27.2
colorful==0.5.5
deprecation==2.1.0
distlib==0.3.7
exceptiongroup==1.1.3
fastapi==0.95.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
gpustat==1.1
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kserve==0.11.0
kubernetes==27.2.0
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencensus==0.11.2
opencensus-context==0.1.3
orjson==3.9.4
packaging==23.1
pandas==2.0.3
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prometheus-client==0.13.1
protobuf==3.20.3
psutil==5.9.5
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.10
pytz==2023.3
PyYAML==6.0.1
ray==2.4.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
timing-asgi==0.3.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
tritonclient==2.36.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.19.0
uvloop==0.17.0
virtualenv==20.21.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
Pillow==10.0.0
